Centralize link qr code UI
