+++ app/src/main/java/im/molly/unifiedpush/components/settings/app/notifications/MollySocketProvisioningQrFragment.kt
@@ -8,23 +8,6 @@ package im.molly.unifiedpush.components.settings.app.notifications
 import android.os.Bundle
 import android.view.View
 import android.view.WindowManager
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.ExperimentalLayoutApi
-import androidx.compose.foundation.layout.FlowRow
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.aspectRatio
-import androidx.compose.foundation.layout.fillMaxHeight
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.widthIn
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
@@ -32,10 +15,7 @@ import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.fragment.app.activityViewModels
 import androidx.lifecycle.DefaultLifecycleObserver
@@ -51,13 +31,11 @@ import kotlinx.coroutines.flow.distinctUntilChanged
 import kotlinx.coroutines.flow.mapNotNull
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
-import org.signal.core.ui.compose.Buttons
 import org.signal.core.ui.compose.Dialogs
 import org.thoughtcrime.securesms.R
-import org.thoughtcrime.securesms.components.settings.app.usernamelinks.QrCode
 import org.thoughtcrime.securesms.compose.ComposeFragment
-import org.thoughtcrime.securesms.compose.SignalTheme
 import org.thoughtcrime.securesms.registration.ui.link.LinkProvisioningQrContract
+import org.thoughtcrime.securesms.registration.ui.link.LinkProvisioningQrSection
 import org.thoughtcrime.securesms.registration.ui.link.LinkProvisioningState
 import org.thoughtcrime.securesms.registration.ui.link.RegisterLinkDeviceQrViewModel
 import org.thoughtcrime.securesms.registration.ui.shared.RegistrationScreen
@@ -125,7 +103,6 @@ class MollySocketProvisioningQrFragment : ComposeFragment() {
   }
 }
 
-@OptIn(ExperimentalLayoutApi::class)
 @Composable
 private fun MollySocketProvisioningQrScreen(
   state: LinkProvisioningState,
@@ -145,89 +122,10 @@ private fun MollySocketProvisioningQrScreen(
       }
     }
   ) {
-    FlowRow(
-      horizontalArrangement = Arrangement.spacedBy(space = 48.dp, alignment = Alignment.CenterHorizontally),
-      verticalArrangement = Arrangement.spacedBy(space = 48.dp),
-      modifier = Modifier
-        .fillMaxWidth()
-        .padding(horizontal = 24.dp)
-    ) {
-      Box(
-        modifier = Modifier
-          .widthIn(160.dp, 320.dp)
-          .aspectRatio(1f)
-          .clip(RoundedCornerShape(24.dp))
-          .background(SignalTheme.colors.colorSurface5)
-          .padding(40.dp)
-      ) {
-        SignalTheme(isDarkMode = false) {
-          Box(
-            modifier = Modifier
-              .clip(RoundedCornerShape(12.dp))
-              .background(MaterialTheme.colorScheme.surface)
-              .fillMaxWidth()
-              .fillMaxHeight()
-              .padding(16.dp),
-            contentAlignment = Alignment.Center
-          ) {
-            when (val qrState = state.qrState) {
-              is RegisterLinkDeviceQrViewModel.QrState.Loaded -> {
-                QrCode(
-                  data = qrState.qrData,
-                  foregroundColor = Color(0xFF2449C0),
-                  modifier = Modifier
-                    .fillMaxWidth()
-                    .fillMaxHeight()
-                )
-              }
-
-              RegisterLinkDeviceQrViewModel.QrState.Loading -> {
-                CircularProgressIndicator(modifier = Modifier.size(48.dp))
-              }
-
-              RegisterLinkDeviceQrViewModel.QrState.Scanned,
-              RegisterLinkDeviceQrViewModel.QrState.Failed -> {
-                val text = if (state.qrState is RegisterLinkDeviceQrViewModel.QrState.Scanned) {
-                  "Scanned on primary device"
-                } else {
-                  stringResource(R.string.RestoreViaQr_qr_code_error)
-                }
-
-                Column(
-                  verticalArrangement = Arrangement.Center,
-                  horizontalAlignment = Alignment.CenterHorizontally
-                ) {
-                  Text(
-                    text = text,
-                    textAlign = TextAlign.Center,
-                    style = MaterialTheme.typography.bodySmall,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant
-                  )
-
-                  Spacer(modifier = Modifier.height(8.dp))
-
-                  Buttons.Small(onClick = onRetry) {
-                    Text(text = stringResource(R.string.RestoreViaQr_retry))
-                  }
-                }
-              }
-            }
-          }
-        }
-      }
-
-      Column(
-        modifier = Modifier
-          .align(alignment = Alignment.CenterVertically)
-          .widthIn(160.dp, 320.dp)
-      ) {
-        Text(
-          text = "Scan this QR code with your primary Signal device to approve a MollySocket credential.",
-          style = MaterialTheme.typography.bodyLarge,
-          color = MaterialTheme.colorScheme.onSurfaceVariant
-        )
-      }
-    }
+    LinkProvisioningQrSection(
+      qrState = state.qrState,
+      onRetry = onRetry
+    )
 
     if (state.hasProvisioningError || error != null) {
       Dialogs.SimpleMessageDialog(
